# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "opcoes-ia — carteira",
    "description": "Superfície para a interface própria. Não decide, não dispara execução e não manda ordem. A leitura é auditada como somente leitura; a escrita é limitada a escrituração da carteira (cadastro de ativo e registro de posição que o usuário já tem). Toda sugestão exposta está pendente de revisão humana.",
    "version": "0.2.0"
  },
  "paths": {
    "/ativos": {
      "get": {
        "tags": [
          "carteira"
        ],
        "summary": "Listar Ativos",
        "description": "Ativos cadastrados — o pré-requisito de todo o resto.\n\n`cotacoes`, `opcoes` e `noticias` têm FK para `ativos`: sem cadastro, o\nETL recusa o ticker e registrar posição em ação falha.",
        "operationId": "listar_ativos_ativos_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AtivoResposta"
                  },
                  "type": "array",
                  "title": "Response Listar Ativos Ativos Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "CognitoAccessToken": []
          }
        ]
      },
      "post": {
        "tags": [
          "carteira"
        ],
        "summary": "Cadastrar Ativo",
        "description": "Cadastra ou CORRIGE um ativo.\n\nRegravar o mesmo ticker atualiza a linha em vez de duplicar ou falhar —\nas cotações já coletadas continuam associadas, porque a chave primária\nnão muda.",
        "operationId": "cadastrar_ativo_ativos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AtivoEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AtivoResposta"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CognitoAccessToken": []
          }
        ]
      }
    },
    "/posicoes": {
      "get": {
        "tags": [
          "carteira"
        ],
        "summary": "Listar Posicoes",
        "description": "Posições em aberto (`fechada_em IS NULL`).",
        "operationId": "listar_posicoes_posicoes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PosicaoAbertaResposta"
                  },
                  "type": "array",
                  "title": "Response Listar Posicoes Posicoes Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "CognitoAccessToken": []
          }
        ]
      },
      "post": {
        "tags": [
          "carteira"
        ],
        "summary": "Registrar Posicao",
        "description": "Registra uma posição que o usuário JÁ tem.\n\nIsto é escrituração, não ordem: o sistema não fala com corretora em\nlugar nenhum. Quantidade negativa registra posição lançada — o caso da\nvenda coberta.",
        "operationId": "registrar_posicao_posicoes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PosicaoEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosicaoCriada"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CognitoAccessToken": []
          }
        ]
      }
    },
    "/posicoes/{posicao_id}/encerrar": {
      "post": {
        "tags": [
          "carteira"
        ],
        "summary": "Encerrar Posicao",
        "description": "Marca `fechada_em` com o desfecho. A linha NUNCA é removida: o\nhistórico é o que permite explicar uma decisão passada meses depois.\n\nExige corpo com o motivo — o que também serve de confirmação: encerrar\né irreversível pela interface (não existe reabrir), e um POST vazio\ntornava perda de dado a um clique de distância.",
        "operationId": "encerrar_posicao_posicoes__posicao_id__encerrar_post",
        "parameters": [
          {
            "name": "posicao_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Posicao Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EncerramentoEntrada"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CognitoAccessToken": []
          }
        ]
      }
    },
    "/watchlist": {
      "get": {
        "tags": [
          "carteira"
        ],
        "summary": "Listar Watchlist",
        "description": "Ativos vigiados e o custo que isso impõe ao orçamento diário.",
        "operationId": "listar_watchlist_watchlist_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistResposta"
                }
              }
            }
          }
        },
        "security": [
          {
            "CognitoAccessToken": []
          }
        ]
      },
      "post": {
        "tags": [
          "carteira"
        ],
        "summary": "Adicionar A Watchlist",
        "description": "Passa a vigiar um ativo JÁ CADASTRADO.\n\nVigiar não cadastra: `ativos` é alvo de FK de cotação, opção e notícia,\ne criar o registro aqui exigiria inventar o nome do ativo — o que a\nregra 1 do projeto proíbe.",
        "operationId": "adicionar_a_watchlist_watchlist_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VigiarEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Adicionar A Watchlist Watchlist Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CognitoAccessToken": []
          }
        ]
      }
    },
    "/watchlist/{ticker}/remover": {
      "post": {
        "tags": [
          "carteira"
        ],
        "summary": "Remover Da Watchlist",
        "description": "Sai da varredura e para de consumir orçamento. O cadastro e o\nhistórico permanecem — isto não descadastra o ativo.",
        "operationId": "remover_da_watchlist_watchlist__ticker__remover_post",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CognitoAccessToken": []
          }
        ]
      }
    },
    "/caixa": {
      "get": {
        "tags": [
          "carteira"
        ],
        "summary": "Ler Caixa",
        "operationId": "ler_caixa_caixa_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaixaResposta"
                }
              }
            }
          }
        },
        "security": [
          {
            "CognitoAccessToken": []
          }
        ]
      },
      "post": {
        "tags": [
          "carteira"
        ],
        "summary": "Lancar Caixa",
        "description": "Registra um lançamento. É o que torna uma PUT coberta, coberta:\nsem garantia para honrar o exercício ao strike, `avaliar()` recusa a\noperação em vez de tratá-la como coberta.",
        "operationId": "lancar_caixa_caixa_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaixaEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Lancar Caixa Caixa Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CognitoAccessToken": []
          }
        ]
      }
    },
    "/catalogo": {
      "get": {
        "tags": [
          "carteira"
        ],
        "summary": "Buscar No Catalogo",
        "description": "Procura ativos por código ou nome no catálogo do provedor.\n\nExiste para o cadastro não depender de digitação: nome e CNPJ vêm da\nfonte em vez da memória do usuário. Um dígito trocado no CNPJ raiz\nquebra o vínculo com o dump da CVM, e o calendário de resultados fica\nsilenciosamente vazio para aquele ativo.\n\nCada chamada consome 1 request do orçamento diário — a interface deve\nbuscar no envio, nunca a cada tecla.",
        "operationId": "buscar_no_catalogo_catalogo_get",
        "parameters": [
          {
            "name": "busca",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Busca"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CandidatoResposta"
                  },
                  "title": "Response Buscar No Catalogo Catalogo Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CognitoAccessToken": []
          }
        ]
      }
    },
    "/catalogo/{ticker}/cnpj": {
      "get": {
        "tags": [
          "carteira"
        ],
        "summary": "Cnpj Do Catalogo",
        "description": "Raiz do CNPJ do ativo, para preencher o vínculo com a CVM.",
        "operationId": "cnpj_do_catalogo_catalogo__ticker__cnpj_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Cnpj Do Catalogo Catalogo  Ticker  Cnpj Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CognitoAccessToken": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AtivoEntrada": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "nome": {
            "type": "string",
            "title": "Nome",
            "description": "Obrigatório: o sistema NUNCA deriva nome a partir do ticker"
          },
          "tipo": {
            "type": "string",
            "title": "Tipo",
            "description": "acao | fii | bdr",
            "default": "acao"
          },
          "cnpj_raiz": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cnpj Raiz",
            "description": "8 dígitos, a RAIZ do CNPJ — é o que liga o ativo ao dump da CVM para datas de resultado"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "nome"
        ],
        "title": "AtivoEntrada"
      },
      "AtivoResposta": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "tipo": {
            "type": "string",
            "title": "Tipo"
          },
          "cnpj_raiz": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cnpj Raiz"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "nome",
          "tipo",
          "cnpj_raiz",
          "criado_em"
        ],
        "title": "AtivoResposta"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PosicaoAbertaResposta": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "tipo_ativo": {
            "type": "string",
            "title": "Tipo Ativo"
          },
          "quantidade": {
            "type": "integer",
            "title": "Quantidade"
          },
          "preco_medio": {
            "type": "number",
            "title": "Preco Medio"
          },
          "aberta_em": {
            "type": "string",
            "format": "date-time",
            "title": "Aberta Em"
          },
          "origem": {
            "type": "string",
            "title": "Origem"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ticker",
          "tipo_ativo",
          "quantidade",
          "preco_medio",
          "aberta_em",
          "origem"
        ],
        "title": "PosicaoAbertaResposta"
      },
      "PosicaoCriada": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "executou_ordem": {
            "type": "boolean",
            "title": "Executou Ordem",
            "description": "Sempre False. Registrar posição é espelhar o que já existe na corretora; nada foi enviado a lugar nenhum",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "PosicaoCriada"
      },
      "PosicaoEntrada": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker",
            "description": "Em ACAO é o ticker do ativo (precisa estar cadastrado); em OPCAO é o CÓDIGO da opção, que não é linha em `ativos`"
          },
          "tipo_ativo": {
            "type": "string",
            "title": "Tipo Ativo",
            "description": "ACAO | OPCAO"
          },
          "quantidade": {
            "type": "integer",
            "title": "Quantidade",
            "description": "Negativo = posição LANÇADA (vendida). É assim que uma venda coberta é registrada. Está em OPÇÕES (cada uma sobre 1 ação), não em contratos"
          },
          "preco_medio": {
            "type": "number",
            "title": "Preco Medio",
            "description": "Base de custo — nunca valor de mercado"
          },
          "ticker_objeto": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ticker Objeto",
            "description": "A ação que a opção cobre. Informado, não inferido: derivar do código exigiria interpretar código B3"
          },
          "strike": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strike"
          },
          "vencimento": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vencimento"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "tipo_ativo",
          "quantidade",
          "preco_medio"
        ],
        "title": "PosicaoEntrada"
      },
...
    },
    "securitySchemes": {
      "CognitoAccessToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token Cognito com o escopo configurado da API."
      }
    }
  }
}